TMDM-11826 MDM server should not start if datasources.xml config points to a system db that does not match the current product version
