+++ org.talend.mdm.core/src/com/amalto/core/objects/configurationinfo/assemble/AssembleBuilder.java
@@ -28,11 +28,6 @@ public interface AssembleBuilder {
 	/** 
 	 * This method constructs and assembles a particular part of a Proc.
 	 */
-	public void buildPartMigrateVersion();
-	
-	/** 
-	 * This method constructs and assembles a particular part of a Proc.
-	 */
 	public void buildPartStartEngine();
 	
 	/**
